+++ src/settings.ts
@@ -2,7 +2,7 @@ export const PLUGIN_NAME = 'homebridge-smartthings-windfree-ac';
 export const PLATFORM_NAME = 'SmartThingsWindFreeAC';
 
 export const DEFAULT_BASE_URL = 'https://api.smartthings.com/v1';
-export const SMARTTHINGS_AUTH_BASE_URL = 'https://api.smartthings.com/oauth';
+export const SMARTTHINGS_AUTH_BASE_URL = 'https://api.smartthings.com/v1/oauth';
 export const DEFAULT_REDIRECT_URI = 'https://httpbin.org/get';
 
 export const REQUIRED_CAPABILITIES = [